test: organize tests code
