+++ test/specs/core-features.e2e.ts
@@ -1,0 +1,58 @@
+import { expect } from '@wdio/globals';
+import type { Browser } from 'webdriverio';
+import { ObsidianBrowserCommands } from 'wdio-obsidian-service';
+import { getFileTreeItem } from './test-utils';
+
+declare const browser: Browser & ObsidianBrowserCommands;
+
+describe('Jupyter for Obsidian', function() {
+    it('can open valid notebook', async () => {
+        // Expect the current server status to be idle (or exited)
+        const idleServerStatus = await browser.$('.side-dock-ribbon-action[aria-label="Start Jupyter Server"]');
+        expect(idleServerStatus).toExist();
+
+        // Click on a Jupyter notebook to open it
+        const notebook = await getFileTreeItem("Notebook creation/Valid notebook.ipynb", browser);
+        await expect(notebook).toExist();
+        await notebook.click();
+
+        // Expect the server status to change to starting
+        const startingServerStatus = await browser.$('.side-dock-ribbon-action[aria-label="Jupyter Server is starting"]');
+        await startingServerStatus.waitForExist({ timeout: 2000 });
+
+        // Wait for a tab to open with the notebook
+        const tabHeader = await browser.$('.workspace-tab-header.is-active[aria-label="Valid notebook.ipynb"]');
+        await tabHeader.waitForExist({ timeout: 5000 });
+
+        // Check that the notebook entry in the tree is marked as active
+        expect((await notebook.getAttribute('class')).includes('is-active')).toBe(true);
+
+        // Check that a message is being displayed while the notebook is loading
+        const viewContent = await browser.$('.view-content');
+        expect(viewContent).toExist();
+        const messageContainer = await viewContent.$('.jupyter-message-container');
+        await messageContainer.waitForExist({ timeout: 2000 });
+
+        // Expect Jupyter status to be running after a while
+        const runningServerStatus = await browser.$('.side-dock-ribbon-action[aria-label="Stop Jupyter Server"]');
+        await runningServerStatus.waitForExist({ timeout: 30000 });
+        
+        // Check that the notebook content is loaded
+        const jupyterWebview = await browser.$('webview.jupyter-webview');
+        await jupyterWebview.waitForExist({ timeout: 5000 });
+
+        // Check that the notebook entry in the tree is still marked as active
+        expect((await notebook.getAttribute('class')).includes('is-active')).toBe(true);
+
+        // Stop the Jupyter server
+        await runningServerStatus.click();
+        await idleServerStatus.waitForExist({ timeout: 2000 });
+
+        // Check that the Jupyter webview is no longer present
+        const jupyterWebviewExists = await jupyterWebview.isExisting();
+        expect(jupyterWebviewExists).toBe(false);
+
+        // Check that the notebook entry in the tree is still marked as active
+        expect((await notebook.getAttribute('class')).includes('is-active')).toBe(true);
+    });
+})
